services/orchestrator-api/requirements.txt
fastapi
#gunicorn
uvicorn[standard]

google-cloud-pubsub>=2.21.0
google-cloud-firestore
google-cloud-storage
